requirements.txt
flask
opencc-python-reimplemented
